requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiofiles>=23.2.1
pydantic==2.7.1
pydantic-settings==2.3.1
email-validator==2.1.1

# LangChain
langchain==0.2.5
langchain-core==0.2.9
langchain-openai==0.1.8
langchain-community==0.2.5

# Vector DB
pinecone-client==4.1.0

# Embeddings
sentence-transformers==3.0.1

# Google APIs
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.134.0

# HTTP
requests==2.32.3

# OCR (optional – used in original notebook)
easyocr==1.7.1

# PDF extraction
pypdf>=4.0.0

# Testing
pytest==8.2.2
httpx==0.27.0
